split up command.rs
