+++ lefthk-core/src/config/command.rs
@@ -1,3 +1,4 @@
+use super::{build_execute, Keybind};
 use crate::errors::{LeftError, Result};
 use std::convert::TryFrom;
 
@@ -22,27 +23,4 @@ impl TryFrom<&str> for Command {
             _ => Err(LeftError::CommandNotFound),
         }
     }
-}
-
-fn build_execute(raw: &str) -> Result<Command> {
-    let headless = without_head(raw, "Execute ");
-    Ok(Command::Execute(headless.to_owned()))
-}
-
-fn without_head<'a, 'b>(s: &'a str, head: &'b str) -> &'a str {
-    if !s.starts_with(head) {
-        return s;
-    }
-    &s[head.len()..]
-}
-
-#[derive(Debug, PartialEq, Clone)]
-pub struct Keybind {
-    pub command: Command,
-    pub modifier: Vec<String>,
-    pub key: String,
-}
-
-pub trait Config {
-    fn mapped_bindings(&self) -> Vec<Keybind>;
 }
