+++ lefthk-core/src/config/keybind.rs
@@ -1,0 +1,9 @@
+use super::Command;
+
+#[derive(Debug, PartialEq, Clone)]
+pub struct Keybind {
+    pub command: Command,
+    pub modifier: Vec<String>,
+    pub key: String,
+}
+
